feat:add: mainで2つのホームを選択し、homeを呼び出す
ホームはhome.dartに分割
